feat: register an adyen instance
