--- examples/adyen-credit-card/src/runtime/plugin.ts
@@ -1,5 +1,0 @@
-import { defineNuxtPlugin } from "#app";
-
-export default defineNuxtPlugin(() => {
-  console.log("Plugin injected by my-module!");
-});
